Implement parsing general section (v3)
Implementing this was a huge mess, and I wouldn't be surprised if things were to
change in the future. Basically I had to plan how to do it, implement simple
permutation parsing, test it all, and fix, fix, and once more fix. For the most
part, GeneralSectionV3 is there just for testing purposes, although I did try
to be smart about the design.