Change TimingPoint to a typeclass
It seems that old version of the .osu format don't have all the properties of
timing points. More precisely, old versions seem to have only offset and
milliseconds per beat. Even version 4 doesn't seem to have all of the properties
found in more recent versions. Thus, I've changed TimingPoint to be a typeclass
so there can be different versions of the actual TimingPoint type in addition to
there being a common interface for all versions of timing points.