Rework errors
This commit brings two changes: First, each parser can produce only one error.
This is because in practice, it's not feasible to produce multiple errors in a
sensible way under the current system. This also makes the second change easier.

The second change is that each parse error can now have a textual description.
The motivation for this is simply easier debugging, because NoParse is a very
bad error when parsing the entire (valid) beatmap fails.